# Conflicts:
#	MiamCommon/Source/EditableEllipse.cpp
#	MiamEdit/Source/GraphicSessionManager.cpp

+++ SpatCommon/Source/LabelledMatrixComponent.cpp
@@ -132,10 +132,10 @@ void LabelledMatrixComponent::mouseMove (const MouseEvent& e)
 
             // Detection of concerned in/out labels
             Label* newInputLabel = 0;
-            if (0<=row && row < maxRowsCount)
+            if (0<=row && row < (int)maxRowsCount)
                 newInputLabel = labels[row];
             Label* newOutputLabel = 0;
-            if (0 <= col && col<maxColsCount)
+            if (0 <= col && col < (int)maxColsCount)
                 newOutputLabel = labels[maxColsCount+col];
 
             // Graphical updates for input
@@ -187,13 +187,13 @@ void LabelledMatrixComponent::ReconstructGuiObjects()
     labels.resize(maxColsCount+maxRowsCount);
 
     // Actual creation of sliders and labels
-    for (int i=0 ; i<maxRowsCount ; i++)
+    for (int i=0 ; i<(int)maxRowsCount ; i++)
     {
         // Label on each row
         labels[i] = new Label("Input label " + std::to_string(i), "" + std::to_string(i+1));
         initAndAddLabel(labels[i]);
     }
-    for (int j=0 ; j<maxColsCount ; j++)
+    for (int j=0 ; j<(int)maxColsCount ; j++)
     {
         // Column labels
         labels[maxRowsCount+j] = new Label("Output label " + std::to_string(j), "" + std::to_string(j+1));
@@ -222,12 +222,12 @@ void LabelledMatrixComponent::repositionLabels()
     const int matrixDeltaY = matrixViewport->getViewPositionY();
 
     // Labels dynamic positionning
-    for (int i=0 ; i<maxRowsCount ; i++)
+    for (int i=0 ; i<(int)maxRowsCount ; i++)
     {
         // Label on each row
         labels[i]->setBounds(0, i*matItemH - matrixDeltaY, inLabelsW, matItemH);
     }
-    for (int j=0 ; j<maxColsCount ; j++)
+    for (int j=0 ; j<(int)maxColsCount ; j++)
     {
         // Column labels
         labels[maxRowsCount+j]->setBounds(inLabelsW+j*matItemW - matrixDeltaX, inOutLabelY, matItemW, outLabelsH);
@@ -269,7 +269,7 @@ void LabelledMatrixComponent::setMatrixToIdentity()
     
     // Then, actual setting to identity
     unsigned int smallestDimension = std::min(maxRowsCount, maxColsCount);
-    for (int i=0 ; i<smallestDimension ; i++)
+    for (int i=0 ; i<(int)smallestDimension ; i++)
         // each one will notify (not optimized but that's OK)
         GetMatrixComponent()->SetSliderValue(i, i, 0.0, NotificationType::sendNotification);
 }
